// src/typings/api/trade.d.ts
export type MarginCoinArrayType = {
  sort: number
  rate: string
  coinId: number
  coinName: string
  selected: boolean
  currencySymbol: string
}